refactor ContentTile -> BlogPostTile
