+++ src/lib/components/BlogRoll.svelte
@@ -10,7 +10,7 @@ Display a collection of blog posts is a responsive grid layout.
 -->
 <script lang="ts">
 	import type { BlogPostIndexItem } from '$lib/schemas/blog';
-	import { ContentTile } from '$lib/components';
+	import BlogPostTile from '$lib/components/BlogPostTile.svelte';
 
 	type Props = {
 		posts: BlogPostIndexItem[];
@@ -21,15 +21,7 @@ Display a collection of blog posts is a responsive grid layout.
 
 <div class="blog-roll" data-testid="blog-roll">
 	{#each posts as post (post.id)}
-		<ContentTile
-			ctaLabel="Read article"
-			href="/blog/{post.slug}"
-			mediaSrc={post.feature_image}
-			mediaAlt={post.feature_image_alt}
-			title={post.title}
-			date={post.published_at}
-			description={post.excerpt}
-		/>
+		<BlogPostTile {post} />
 	{:else}
 		<p>No blog posts found (check if Ghost is properly configured)</p>
 	{/each}
