+++ src/lib/components/index.ts
@@ -2,12 +2,12 @@ export { default as Alert } from './Alert.svelte';
 export { default as AlertItem } from './AlertItem.svelte';
 export { default as AlertList } from './AlertList.svelte';
 export { default as Banner } from './Banner.svelte';
+export { default as BlogPostTile } from './BlogPostTile.svelte';
 export { default as BlogRoll } from './BlogRoll.svelte';
 export { default as Button } from './Button.svelte';
 export { default as ContentCard } from './ContentCard.svelte';
 export { default as ContentCardsSection } from './ContentCardsSection.svelte';
 export { default as ContentCardsTemplate } from './ContentCardsTemplate.svelte';
-export { default as ContentTile } from './ContentTile.svelte';
 export { default as CopyWidget } from './CopyWidget.svelte';
 export { default as CryptoAddressWidget } from './CryptoAddressWidget.svelte';
 export { default as DataBadge } from './DataBadge.svelte';
